This change addresses all warnings which show up with terrafrom 0.12+. Below is one such warning. A total of 144 warnings were eliminated. Terraform versions prior to 0.12 will show errors for new HCL2 constructs.
Warning: Interpolation-only expressions are deprecated

  on /home/ubuntu/dev/terraform-aws-cognito-auth/modules/api/main.tf line 44, in resource "aws_iam_role" "lambda":
  44:   assume_role_policy = "${
  45:     file("${path.module}/iam/policies/assume-role/lambda.json")
  46:   }"

Terraform 0.11 and earlier required all non-constant expressions to be
provided via interpolation syntax, but this pattern is now deprecated. To
silence this warning, remove the "${ sequence from the start and the }"
sequence from the end of this expression, leaving just the inner expression.

Template interpolation syntax is still used to construct strings from
expressions when the template includes multiple interpolation sequences or a
mixture of literal strings and interpolations. This deprecation applies only
to templates that consist entirely of a single interpolation sequence.

(and 136 more similar warnings elsewhere)